@@ -11,51 +11,52 @@
     "postinstall": "prisma generate"
   },
   "dependencies": {
-    "@auth/prisma-adapter": "^2.8.0",
-    "@hookform/resolvers": "^5.0.1",
-    "@prisma/client": "^6.6.0",
-    "@radix-ui/react-avatar": "^1.1.6",
+    "@auth/prisma-adapter": "^2.11.1",
+    "@hookform/resolvers": "^5.2.2",
+    "@prisma/adapter-mariadb": "^7.2.0",
+    "@prisma/client": "^7.2.0",
+    "@radix-ui/react-avatar": "^1.1.11",
     "@radix-ui/react-checkbox": "^1.3.3",
-    "@radix-ui/react-dialog": "^1.1.7",
-    "@radix-ui/react-dropdown-menu": "^2.1.7",
+    "@radix-ui/react-dialog": "^1.1.15",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
     "@radix-ui/react-icons": "^1.3.2",
-    "@radix-ui/react-label": "^2.1.3",
-    "@radix-ui/react-popover": "^1.1.7",
-    "@radix-ui/react-scroll-area": "^1.2.4",
-    "@radix-ui/react-select": "^2.1.7",
-    "@radix-ui/react-separator": "^1.1.3",
-    "@radix-ui/react-slot": "^1.2.0",
-    "@tanstack/react-table": "^8.21.2",
+    "@radix-ui/react-label": "^2.1.8",
+    "@radix-ui/react-popover": "^1.1.15",
+    "@radix-ui/react-scroll-area": "^1.2.10",
+    "@radix-ui/react-select": "^2.2.6",
+    "@radix-ui/react-separator": "^1.1.8",
+    "@radix-ui/react-slot": "^1.2.4",
+    "@tanstack/react-table": "^8.21.3",
     "@types/react-signature-canvas": "^1.0.7",
-    "@types/uuid": "^10.0.0",
-    "basic-ftp": "^5.0.5",
+    "basic-ftp": "^5.1.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "date-fns": "^4.1.0",
-    "lucide-react": "^0.487.0",
-    "next": "15.3.8",
-    "next-auth": "^4.24.11",
-    "react": "^19.0.0",
-    "react-day-picker": "^9.6.7",
-    "react-dom": "^19.0.0",
-    "react-hook-form": "^7.55.0",
+    "lucide-react": "^0.562.0",
+    "mariadb": "^3.4.5",
+    "next": "16.1.1",
+    "next-auth": "^4.24.13",
+    "react": "^19.2.3",
+    "react-day-picker": "^9.13.0",
+    "react-dom": "^19.2.3",
+    "react-hook-form": "^7.70.0",
     "react-signature-canvas": "^1.1.0-alpha.2",
-    "simple-icons": "^14.12.1",
-    "sonner": "^2.0.3",
-    "tailwind-merge": "^3.2.0",
-    "tw-animate-css": "^1.2.5",
-    "uuid": "^11.1.0",
-    "zod": "^3.24.2"
+    "simple-icons": "^16.4.0",
+    "sonner": "^2.0.7",
+    "tailwind-merge": "^3.4.0",
+    "tw-animate-css": "^1.4.0",
+    "uuid": "^13.0.0",
+    "zod": "^4.3.5"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4",
-    "@types/node": "20.19.20",
-    "@types/react": "19.2.2",
+    "@types/node": "25.0.3",
+    "@types/react": "19.2.7",
     "@types/react-dom": "^19",
-    "prisma": "^6.6.0",
+    "prisma": "^7.2.0",
     "tailwindcss": "^4",
-    "tsx": "^4.19.3",
+    "tsx": "^4.21.0",
     "typescript": "5.9.3"
   }
 }
