@@ -2,4 +2,4 @@
 /* !!! This is code generated by Prisma. Do not edit directly. !!!
 /* eslint-disable */
 // biome-ignore-all lint: generated file
-module.exports = { ...require('.') }
+export default import('./query_compiler_bg.wasm?module')
